+++ paddlenlp/utils/import_utils.py
@@ -13,16 +13,96 @@
 # limitations under the License.
 from __future__ import annotations
 
+import builtins
+import functools
 import importlib.util
 import os
 import shutil
 import site
 import sys
+from contextlib import contextmanager
 from typing import Optional, Tuple, Type, Union
 
 import pip
 
 from paddlenlp.utils.log import logger
+
+_original_import = builtins.__import__
+_imported_modules = {}
+_paddlenlp_ops_updated = False
+_original_attributes = {}
+pybind_ops_list = [
+    "update_inputs_v2",
+    "save_output",
+    "set_preids_token_penalty_multi_scores",
+    "rebuild_padding_v2",
+    "append_attention",
+    "save_output_dygraph",
+]
+
+
+def custom_import(name, *args, **kwargs):
+    global _paddlenlp_ops_updated, _imported_modules, _original_attributes
+    global pybind_ops_list
+
+    if _paddlenlp_ops_updated:
+        if name in _imported_modules:
+            return _imported_modules[name]
+
+    module = _original_import(name, *args, **kwargs)
+
+    if not _paddlenlp_ops_updated and os.getenv("USE_PYBIND_CUSTOM_OPS", "1").lower() in [
+        "1",
+        "true",
+        "t",
+        "yes",
+        "y",
+    ]:
+        if name == "paddlenlp_ops":
+            logger.info("Using Pybind paddlenlp_ops!")
+            if name not in _original_attributes:
+                bak_dict = {}
+                for ops_name in pybind_ops_list:
+                    bak_dict[ops_name] = getattr(module, ops_name, None)
+                _original_attributes[name] = bak_dict
+
+            for ops_name in pybind_ops_list:
+                pybind_ops_name = f"f_{ops_name}"
+                if hasattr(module, pybind_ops_name):
+                    setattr(module, ops_name, getattr(module, pybind_ops_name))
+
+            _paddlenlp_ops_updated = True
+
+    _imported_modules[name] = module
+    return module
+
+
+@contextmanager
+def dynamic_graph_pybind_context():
+    global _original_import, _paddlenlp_ops_updated
+    original_import = builtins.__import__
+
+    try:
+        builtins.__import__ = custom_import
+        yield
+    finally:
+        builtins.__import__ = original_import
+
+        if "paddlenlp_ops" in _original_attributes:
+            paddlenlp_ops_module = sys.modules.get("paddlenlp_ops")
+            if paddlenlp_ops_module:
+                for attr, value in _original_attributes["paddlenlp_ops"].items():
+                    setattr(paddlenlp_ops_module, attr, value)
+                _paddlenlp_ops_updated = False
+
+
+def auto_dynamic_graph_pybind(func):
+    @functools.wraps(func)
+    def wrapper(self, *args, **kwargs):
+        with dynamic_graph_pybind_context():
+            return func(self, *args, **kwargs)
+
+    return wrapper
 
 
 # TODO: This doesn't work for all packages (`bs4`, `faiss`, etc.) Talk to Sylvain to see how to do with it better.
